refactor: split e2e-support lib modules and ratchet the 250-line cap (#4598)
Topic-split remaining over-cap helpers so wave-4 Slice A can enforce an empty
250-line lib allowlist without changing poll ramps or public AC1 names.
