+++ packages/colonizethis_app_e2e_support/lib/e2e_test_shared_first_fleet_move.dart
@@ -6,47 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app_e2e_support/e2e_test_shared.dart';
 
-/// Outcome of one [e2eAttemptFirstFleetMoveOrCancel] invocation.
-///
-/// The full-turn E2E scenario (`new_game_full_turn_e2e_test.dart`) inlined
-/// the post-split "tap first Move; cancel if no destinations" block until
-/// this lift (Refs GitHub #2336 AC1 / AC2). Exposing the three outcomes as
-/// distinct enum values lets call sites attribute wall-clock segments
-/// without re-decoding which branch fired (replaces ad-hoc `if`
-/// branching on the helper's side effects).
-enum E2eFirstFleetMoveOutcome {
-  /// No keyed Move button ([kCtE2EFleetMoveActionKey]) descendant of
-  /// [kCtE2ENavalPanelRootKey] was found; no dialog was opened.
-  noMoveButton,
-
-  /// The move dialog opened but contained no `RadioListTile<dynamic>`
-  /// destinations; the helper tapped Cancel and waited until the dialog
-  /// dismissed.
-  cancelled,
-
-  /// The move dialog opened, a destination was tapped, Confirm was tapped,
-  /// and the dialog dismissed within the budget.
-  confirmed,
-}
-
-/// Default cap for `wait_until_move_dialog_after_tap` inside
-/// [e2eAttemptFirstFleetMoveOrCancel]. Matches the legacy 5 s `waitUntilFound`
-/// timeout the inline full-turn block used (Refs GitHub #2336 AC1 / AC2).
-const Duration kE2eDefaultFirstFleetMoveDialogOpenTimeout = Duration(seconds: 5);
-
-/// Default cap for `pump_until_move_confirm_tappable` inside
-/// [e2eAttemptFirstFleetMoveOrCancel]. Matches the legacy 2 s adaptive wait
-/// the inline full-turn block used after tapping a destination radio
-/// (Refs GitHub #2336 AC1 / AC2).
-const Duration kE2eDefaultFirstFleetMoveConfirmReadyTimeout =
-    Duration(seconds: 2);
-
-/// Default cap for `pump_until_move_dialog_closed*` inside
-/// [e2eAttemptFirstFleetMoveOrCancel]. Matches the legacy 10 s cap the
-/// inline full-turn block used for both the confirm and cancel paths
-/// (Refs GitHub #2336 AC1 / AC2).
-const Duration kE2eDefaultFirstFleetMoveDialogCloseTimeout =
-    Duration(seconds: 10);
+export 'e2e_test_shared_first_fleet_move_types.dart';
 
 /// Opportunistically taps the first keyed Move action descendant of the open
 /// naval panel, picks the first destination radio in the resulting
@@ -167,10 +127,7 @@ Future<E2eFirstFleetMoveOutcome> e2eAttemptFirstFleetMoveOrCancel(
   );
   if (destinationRadios.evaluate().isEmpty) {
     final cancel = find
-        .descendant(
-          of: moveDialog,
-          matching: find.text(l10n.common_cancel),
-        )
+        .descendant(of: moveDialog, matching: find.text(l10n.common_cancel))
         .hitTestable();
     expect(
       cancel,
@@ -199,10 +156,7 @@ Future<E2eFirstFleetMoveOutcome> e2eAttemptFirstFleetMoveOrCancel(
   await e2ePumpUntilConditionOrIdle(
     tester,
     () => find
-        .descendant(
-          of: moveDialog,
-          matching: find.text(l10n.common_confirm),
-        )
+        .descendant(of: moveDialog, matching: find.text(l10n.common_confirm))
         .hitTestable()
         .evaluate()
         .isNotEmpty,
@@ -211,10 +165,7 @@ Future<E2eFirstFleetMoveOutcome> e2eAttemptFirstFleetMoveOrCancel(
     phaseName: 'pump_until_move_confirm_tappable',
   );
   final confirm = find
-      .descendant(
-        of: moveDialog,
-        matching: find.text(l10n.common_confirm),
-      )
+      .descendant(of: moveDialog, matching: find.text(l10n.common_confirm))
       .hitTestable();
   expect(
     confirm,
