+++ packages/colonizethis_app_e2e_support/lib/e2e_test_shared_fleet_reach_nw_predicates.dart
@@ -1,5 +1,6 @@
 // Fleet-reach New World predicates — topic-split aggregator (Refs #4195 slice A).
 export 'e2e_test_shared_fleet_reach_nw_civilian.dart';
+export 'e2e_test_shared_fleet_reach_nw_coastal.dart';
 export 'e2e_test_shared_fleet_reach_nw_panel.dart';
 export 'e2e_test_shared_fleet_reach_nw_snapshot.dart';
 export 'e2e_test_shared_fleet_reach_nw_text.dart';
