+++ packages/colonizethis_app_e2e_support/lib/test_support/production_panel_e2e_expected_lines.dart
@@ -7,173 +7,13 @@
 // Expected plain-text lines for ProductionPanel (wide layout ≥ kNarrowBreakpoint).
 // Mirrors app/lib/features/game/widgets/production/production_panel.dart for e2e.
 
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_app_fixtures/config/ct_e2e_last_panel_snapshot.dart';
-import 'package:colonizethis_app/features/game/widgets/production/production_recipe_affordance.dart';
 import 'package:colonizethis_app/features/game/widgets/production/production_labour_helpers.dart';
 import 'package:colonizethis_app_l10n/l10n/l10n.dart';
-import 'package:colonizethis_app/widgets/ct_resource_cell.dart';
-
-Set<String> _inputCommodityIds() {
-  final inputIds = <String>{};
-  for (final recipe in ProductionRecipesCatalog.all) {
-    inputIds.addAll(recipe.inputQuantities.keys);
-  }
-  return inputIds;
-}
-
-String _workerDisplayName(String workerType, AppLocalizations l10n) {
-  switch (workerType) {
-    case 'peasant':
-      return l10n.production_workers_peasants;
-    case 'apprentice':
-      return l10n.production_workers_apprentices;
-    case 'journeyman':
-      return l10n.production_workers_journeymen;
-    case 'master':
-      return l10n.production_workers_masters;
-    default:
-      return workerType;
-  }
-}
-
-void _addAvailableTexts(
-  List<String> out,
-  CtE2eProductionPanelSnapshot snap,
-  AppLocalizations l10n,
-  int effectiveLabour,
-) {
-  final player = snap.player;
-  final netChanges = snap.netDeltasByCommodity;
-  final inputCommodityIds = _inputCommodityIds();
-
-  final rawMaterials = CommodityCatalog.all
-      .where(
-        (c) =>
-            c.category == CommodityCategory.rawMaterial &&
-            inputCommodityIds.contains(c.id),
-      )
-      .toList();
-  final manufactured = CommodityCatalog.all
-      .where((c) => c.category == CommodityCategory.manufactured)
-      .toList();
-  final availableFood = CommodityCatalog.all
-      .where((c) => c.category == CommodityCategory.food)
-      .toList();
-
-  out.add(l10n.production_available);
-  out.add(l10n.production_breakdown);
-
-  void addCommodityCellTexts(Commodity c) {
-    final name = c.displayName ?? c.id;
-    final qty = player.stockpile.quantityOf(c.id);
-    final change = netChanges[c.id] ?? 0;
-    out.add(name);
-    out.add(CtResourceCell.formatQuantity(qty));
-    if (change != 0) {
-      out.add(CtResourceCell.formattedDeltaText(change)!);
-    }
-  }
-
-  void addWorkerCellTexts(String workerType, int count) {
-    out.add(_workerDisplayName(workerType, l10n));
-    out.add(CtResourceCell.formatQuantity(count));
-  }
-
-  if (availableFood.isNotEmpty) {
-    out.add(l10n.production_food.toUpperCase());
-    for (final c in availableFood) {
-      addCommodityCellTexts(c);
-    }
-  }
-
-  out.add(l10n.production_rawMaterials.toUpperCase());
-  for (final c in rawMaterials) {
-    addCommodityCellTexts(c);
-  }
-
-  if (manufactured.isNotEmpty) {
-    out.add(l10n.production_manufactured.toUpperCase());
-    for (final c in manufactured) {
-      addCommodityCellTexts(c);
-    }
-  }
-
-  out.add(l10n.production_workers.toUpperCase());
-  addWorkerCellTexts('peasant', player.workerPool.peasants);
-  addWorkerCellTexts('apprentice', player.workerPool.apprentices);
-  addWorkerCellTexts('journeyman', player.workerPool.journeymen);
-  addWorkerCellTexts('master', player.workerPool.masters);
-  out.add(l10n.production_labourThisTurn(effectiveLabour));
-}
-
-String _recipeLabelText(ProductionRecipe recipe) {
-  final outputCommodity = CommodityCatalog.byId[recipe.outputCommodityId];
-  final outputName = outputCommodity?.displayName ?? recipe.outputCommodityId;
-  final inputParts = recipe.inputQuantities.entries
-      .map((e) {
-        final comm = CommodityCatalog.byId[e.key];
-        final name = comm?.displayName ?? e.key;
-        return '$name ×${e.value}';
-      })
-      .join(', ');
-  return '$outputName ($inputParts)';
-}
-
-void _addAllocationTexts(
-  List<String> out,
-  CtE2eProductionPanelSnapshot snap,
-  AppLocalizations l10n,
-  int effectiveLabour,
-) {
-  final player = snap.player;
-  final desiredOutputByRecipe = snap.desiredOutputByRecipe;
-
-  var totalRequiredLabour = 0;
-  for (final e in desiredOutputByRecipe.entries) {
-    final recipe = ProductionRecipesCatalog.byId[e.key];
-    if (recipe == null) continue;
-    totalRequiredLabour += e.value * recipe.labourPerOutput;
-  }
-  final labourInsufficient = totalRequiredLabour > effectiveLabour;
-
-  out.add(l10n.production_allocation);
-  out.add(l10n.common_reset);
-
-  for (final recipe in ProductionRecipesCatalog.all) {
-    final desired = desiredOutputByRecipe[recipe.id] ?? 0;
-    final locked = !ProductionRecipesCatalog.isRecipeAvailableForPlayer(
-      recipe,
-      player.techUnlocked,
-    );
-    final affordance = computeRecipeAffordance(
-      recipe: recipe,
-      stockpile: player.stockpile,
-      desiredOutputByRecipe: desiredOutputByRecipe,
-      effectiveLabour: effectiveLabour,
-    );
-    // A locked (tech-gated) recipe forces maxAchievable to 0 and renders the
-    // localized (locked) marker as a separate Text after the recipe label,
-    // mirroring ProductionAllocationRow / _AllocationSubpanel._buildRecipeLabel.
-    final maxAchievable = locked ? 0 : affordance.maxDesiredOutput;
-
-    out.add(_recipeLabelText(recipe));
-    if (locked) {
-      out.add(l10n.production_recipeLocked);
-    }
-    out.add(
-      l10n.production_recipeAffordance(maxAchievable, affordance.limitingLabel),
-    );
-    out.add(desired.toString());
-  }
-
-  out.add(l10n.production_totalLabour(totalRequiredLabour, effectiveLabour));
-  if (labourInsufficient) {
-    out.add(l10n.production_labourInsufficient);
-  }
-}
+import 'production_panel_e2e_expected_lines_allocation.dart';
+import 'production_panel_e2e_expected_lines_available.dart';
 
 String _labourTierDisplayName(WorkerTier tier, AppLocalizations l10n) {
   switch (tier) {
@@ -261,7 +101,7 @@ List<String> productionPanelWideExpectedTexts(
   final effectiveLabour = labourReadiness.effectiveLabour;
 
   final out = <String>[];
-  _addAvailableTexts(out, snap, l10n, effectiveLabour);
+  addProductionPanelAvailableTexts(out, snap, l10n, effectiveLabour);
   out.addAll(
     productionLabourControlsExpectedTexts(
       player: snap.player,
@@ -270,6 +110,6 @@ List<String> productionPanelWideExpectedTexts(
       l10n: l10n,
     ),
   );
-  _addAllocationTexts(out, snap, l10n, effectiveLabour);
+  addProductionPanelAllocationTexts(out, snap, l10n, effectiveLabour);
   return out;
 }
