+++ packages/colonizethis_app_e2e_support/lib/test_support/province_panel_e2e_expected_lines_road.dart
@@ -1,0 +1,19 @@
+// coverage:ignore-file
+// E2E expected-line helpers for province overlay road/rail transport labels.
+
+const String kRoadRailPrimitiveVersusRailGloss =
+    'Basic land link for connectivity and yield caps. Railroads are transport level 4.';
+
+String roadRailSupplementaryLabel(int roadLevel) {
+  return switch (roadLevel) {
+    0 => 'none',
+    1 => 'primitive road',
+    2 => 'improved road',
+    4 => 'port or railroad',
+    _ => 'non-standard transport level',
+  };
+}
+
+String roadRailTransportLevelPrimaryLine(int transportLevel) {
+  return 'Road / railroad: transport level $transportLevel';
+}
