+++ packages/colonizethis_app_e2e_support/test/e2e_shared_lib_topic_split_test.dart
@@ -47,7 +47,7 @@ void main() {
           continue;
         }
         final lines = file.readAsLinesSync().length;
-        if (lines > 300) {
+        if (lines > 250) {
           oversized.add('$name ($lines)');
         }
       }
@@ -56,7 +56,9 @@ void main() {
   });
 
   group('e2e_test_shared_panels topic split (Refs #4075 AC1 / AC2)', () {
-    test('panels umbrella exports production/fleet/explore topic libraries', () {
+    test(
+      'panels umbrella exports production/fleet/explore topic libraries',
+      () {
         final barrel = File(
           p.join(libDir.path, 'e2e_test_shared_panels.dart'),
         ).readAsStringSync();
@@ -109,62 +111,85 @@ void main() {
         final file = File(p.join(libDir.path, name));
         expect(file.existsSync(), isTrue, reason: '$name missing');
         final lines = file.readAsLinesSync().length;
-        expect(lines, lessThanOrEqualTo(300), reason: '$name still over wave-3 300 cap');
+        expect(
+          lines,
+          lessThanOrEqualTo(300),
+          reason: '$name still over wave-3 300 cap',
+        );
       }
     });
 
-    test('slice B residual lib files stay under 300 physical lines (Refs #4195)', () {
-      for (final name in <String>[
-        'e2e_test_shared.dart',
-        'e2e_test_shared_adaptive_polling.dart',
-        'e2e_test_shared_explore_assign.dart',
-        'e2e_test_shared_explore_assign_sweep.dart',
-        'e2e_test_shared_explore_assign_bundled.dart',
-        'test_support/civilian_units_panel_e2e_expected_lines.dart',
-        'test_support/civilian_units_panel_e2e_expected_lines_assigned.dart',
-        'test_support/civilian_units_panel_e2e_expected_lines_rows.dart',
-      ]) {
-        final file = File(p.join(libDir.path, name));
-        expect(file.existsSync(), isTrue, reason: '$name missing');
-        final lines = file.readAsLinesSync().length;
-        expect(lines, lessThanOrEqualTo(300), reason: '$name still over wave-3 300 cap');
-      }
-    });
-
-    test('explore assign topic split exports sibling modules (Refs #4195 slice B)', () {
-      final barrel = File(
-        p.join(libDir.path, 'e2e_test_shared_explore_assign.dart'),
-      ).readAsStringSync();
-      for (final export in <String>[
-        "export 'e2e_test_shared_explore_assign_sweep.dart';",
-        "export 'e2e_test_shared_explore_assign_bundled.dart';",
-      ]) {
-        expect(barrel, contains(export));
-      }
-    });
-
-    test('shared umbrella exports adaptive polling sibling (Refs #4195 slice B)', () {
-      final barrel = File(
-        p.join(libDir.path, 'e2e_test_shared.dart'),
-      ).readAsStringSync();
-      expect(
-        barrel,
-        contains("export 'e2e_test_shared_adaptive_polling.dart';"),
-      );
-    });
-
-    test('slice D test mirror files stay under 400 physical lines (Refs #4344)', () {
-      final testDir = Directory(p.join(e2eSupportPackageRoot().path, 'test'));
-      final oversized = <String>[];
-      for (final file in testDir.listSync(recursive: true).whereType<File>()) {
-        if (!file.path.endsWith('.dart')) continue;
-        final lines = file.readAsLinesSync().length;
-        if (lines > 400) {
-          oversized.add('${p.relative(file.path, from: testDir.path)} ($lines)');
+    test(
+      'slice B residual lib files stay under 300 physical lines (Refs #4195)',
+      () {
+        for (final name in <String>[
+          'e2e_test_shared.dart',
+          'e2e_test_shared_adaptive_polling.dart',
+          'e2e_test_shared_explore_assign.dart',
+          'e2e_test_shared_explore_assign_sweep.dart',
+          'e2e_test_shared_explore_assign_bundled.dart',
+          'test_support/civilian_units_panel_e2e_expected_lines.dart',
+          'test_support/civilian_units_panel_e2e_expected_lines_assigned.dart',
+          'test_support/civilian_units_panel_e2e_expected_lines_rows.dart',
+        ]) {
+          final file = File(p.join(libDir.path, name));
+          expect(file.existsSync(), isTrue, reason: '$name missing');
+          final lines = file.readAsLinesSync().length;
+          expect(
+            lines,
+            lessThanOrEqualTo(300),
+            reason: '$name still over wave-3 300 cap',
+          );
         }
-      }
-      expect(oversized, isEmpty, reason: 'All test files must be ≤500 lines');
-    });
+      },
+    );
+
+    test(
+      'explore assign topic split exports sibling modules (Refs #4195 slice B)',
+      () {
+        final barrel = File(
+          p.join(libDir.path, 'e2e_test_shared_explore_assign.dart'),
+        ).readAsStringSync();
+        for (final export in <String>[
+          "export 'e2e_test_shared_explore_assign_sweep.dart';",
+          "export 'e2e_test_shared_explore_assign_bundled.dart';",
+        ]) {
+          expect(barrel, contains(export));
+        }
+      },
+    );
+
+    test(
+      'shared umbrella exports adaptive polling sibling (Refs #4195 slice B)',
+      () {
+        final barrel = File(
+          p.join(libDir.path, 'e2e_test_shared.dart'),
+        ).readAsStringSync();
+        expect(
+          barrel,
+          contains("export 'e2e_test_shared_adaptive_polling.dart';"),
+        );
+      },
+    );
+
+    test(
+      'slice D test mirror files stay under 400 physical lines (Refs #4344)',
+      () {
+        final testDir = Directory(p.join(e2eSupportPackageRoot().path, 'test'));
+        final oversized = <String>[];
+        for (final file
+            in testDir.listSync(recursive: true).whereType<File>()) {
+          if (!file.path.endsWith('.dart')) continue;
+          final lines = file.readAsLinesSync().length;
+          if (lines > 400) {
+            oversized.add(
+              '${p.relative(file.path, from: testDir.path)} ($lines)',
+            );
+          }
+        }
+        expect(oversized, isEmpty, reason: 'All test files must be ≤500 lines');
+      },
+    );
 
     test('wave-3 Slice B hit_testable_scroll topic split (Refs #4344 AC3)', () {
       final barrel = File(
@@ -192,18 +217,21 @@ void main() {
       }
     });
 
-    test('wave-3 Slice B standard_scenario_opener topic split (Refs #4344 AC3)', () {
-      final barrel = File(
-        p.join(libDir.path, 'e2e_test_shared_standard_scenario_opener.dart'),
-      ).readAsStringSync();
-      for (final export in <String>[
-        "export 'e2e_test_shared_standard_scenario_opener_constants.dart';",
-        "export 'e2e_test_shared_standard_scenario_opener_result.dart';",
-        "export 'e2e_test_shared_standard_scenario_opener_enter.dart';",
-      ]) {
-        expect(barrel, contains(export));
-      }
-    });
+    test(
+      'wave-3 Slice B standard_scenario_opener topic split (Refs #4344 AC3)',
+      () {
+        final barrel = File(
+          p.join(libDir.path, 'e2e_test_shared_standard_scenario_opener.dart'),
+        ).readAsStringSync();
+        for (final export in <String>[
+          "export 'e2e_test_shared_standard_scenario_opener_constants.dart';",
+          "export 'e2e_test_shared_standard_scenario_opener_result.dart';",
+          "export 'e2e_test_shared_standard_scenario_opener_enter.dart';",
+        ]) {
+          expect(barrel, contains(export));
+        }
+      },
+    );
 
     test('wave-3 Slice B naval_move_pick topic split (Refs #4344 AC3)', () {
       final barrel = File(
@@ -242,12 +270,42 @@ void main() {
       }
     });
 
-    test('wave-3 Slice B aliases_ui topic split (Refs #4344)', () {
+    test('wave-4 Slice A topic splits stay ≤250 (Refs #4598)', () {
       final barrel = File(
-        p.join(libDir.path, 'e2e_helpers_aliases_ui.dart'),
+        p.join(libDir.path, 'e2e_test_shared_adaptive_polling.dart'),
       ).readAsStringSync();
-      expect(barrel, contains("export 'e2e_helpers_aliases_ui_wait.dart';"));
-      expect(barrel, contains("export 'e2e_helpers_aliases_ui_panels.dart';"));
+      expect(
+        barrel,
+        contains("export 'e2e_test_shared_adaptive_polling_core.dart';"),
+      );
+      expect(
+        barrel,
+        contains("export 'e2e_test_shared_adaptive_polling_waits.dart';"),
+      );
+      for (final name in <String>[
+        'e2e_test_shared_adaptive_polling.dart',
+        'e2e_test_shared_adaptive_polling_core.dart',
+        'e2e_test_shared_adaptive_polling_waits.dart',
+        'e2e_test_shared_fleet_reach_loop.dart',
+        'e2e_test_shared_fleet_reach_loop_types.dart',
+        'e2e_test_shared_first_fleet_move.dart',
+        'e2e_test_shared_first_fleet_move_types.dart',
+        'e2e_test_shared_fleet_reach_nw_snapshot.dart',
+        'e2e_test_shared_fleet_reach_nw_coastal.dart',
+        'e2e_test_shared_panel_open_outer_loop.dart',
+        'e2e_test_shared_panel_open_outer_loop_body.dart',
+        'e2e_helpers_aliases_orders.dart',
+        'test_support/production_panel_e2e_expected_lines.dart',
+        'test_support/production_panel_e2e_expected_lines_available.dart',
+        'test_support/production_panel_e2e_expected_lines_allocation.dart',
+        'test_support/province_panel_e2e_expected_lines_ctx.dart',
+        'test_support/province_panel_e2e_expected_lines_road.dart',
+      ]) {
+        final file = File(p.join(libDir.path, name));
+        expect(file.existsSync(), isTrue, reason: '$name missing');
+        final lines = file.readAsLinesSync().length;
+        expect(lines, lessThanOrEqualTo(250), reason: '$name still over 250');
+      }
     });
   });
 }
